Revert "0.1.51"

+++ coston/protocol/interfaces/IIFlareSystemsCalculator.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.7.6 <0.9;
 
-import {IFlareSystemsCalculator} from "../../IFlareSystemsCalculator.sol";
+import "../../IFlareSystemsCalculator.sol";
 
 /**
  * FlareSystemsCalculator internal interface.
@@ -18,7 +18,7 @@ interface IIFlareSystemsCalculator is IFlareSystemsCalculator {
      */
     function calculateRegistrationWeight(
         address _voter,
-        uint32 _rewardEpochId,
+        uint24 _rewardEpochId,
         uint256 _votePowerBlockNumber
     ) external returns (uint256 _registrationWeight);
